requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.0
Django==5.1.4
django-credit-cards==0.4.1
django-parler==2.3
django-parler-rest==2.2
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
parler==1.0.1
phonenumbers==8.13.52
PyJWT==2.10.1
requests==2.32.3
sqlparse==0.5.3
urllib3==2.2.3
